Change `Object.LastModified` to a conditionally set pointer
It turns out that although `Last-Modified` is set on the vast majority
of WaniKani responses, it's not set on _all_. One example of this case
is if we request a list filtered for specific IDs.

Here we make `Object.LastModified` a pointer to represent the fact that
it may not always have a usable value.